pak dis shit double

+++ src/model/Product.java
@@ -4,9 +4,9 @@ public class Product
 {
 	private int prod_id;
 	private String name;
-	private float price;
+	private double price;
 	private String category;
-	public Product(int prod_id, String name, float price, String category) 
+	public Product(int prod_id, String name, double price, String category) 
 	{
 		super();
 		this.prod_id = prod_id;
@@ -15,38 +15,38 @@ public class Product
 		this.category = category;
 	}
 	
-	public int getProd_id() 
+	public int getProd_id()
 	{
 		return prod_id;
 	}
-	public void setProd_id(int prod_id) 
+	public void setProd_id(int prod_id)
 	{
 		this.prod_id = prod_id;
 	}
 	
-	public String getName() 
+	public String getName()
 	{
 		return name;
 	}
-	public void setName(String name) 
+	public void setName(String name)
 	{
 		this.name = name;
 	}
 	
-	public float getPrice() 
+	public double getPrice()
 	{
 		return price;
 	}
-	public void setPrice(float price) 
+	public void setPrice(double price)
 	{
 		this.price = price;
 	}
 	
-	public String getCategory() 
+	public String getCategory()
 	{
 		return category;
 	}
-	public void setCategory(String category) 
+	public void setCategory(String category)
 	{
 		this.category = category;
 	}
